# restAPI/src/main/swagger/openapi.yaml
openapi: 3.0.0
info:
  description: This is the API for the UI's access to data storage
  version: 1.0.0
  title: AIDA Store REST API
servers:
  - url: "{protocol}://{host}:{port}/{basePath}"
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: "8008"
      basePath:
        default: api
...
paths:
  /graphs:
    get:
      tags:
        - Graph
      summary: Get a list of named graphs
      description: Get a list of available graphs
      operationId: graph
      responses:
        "200":
          description: successful operation
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
              description: CORS access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/graphs"
        "400":
          description: Invalid
        "404":
          description: API not found
...
  /graph:
    post:
      tags:
        - Blazegraph
      summary: Upload a single TTL and Import into Blazegraph
      description: Upload a single TTL and Import into Blazegraph
      operationId: graphUpload
      parameters:
        - name: graphURI
          in: query
          description: URI of graph
          required: true
          schema:
            type: string
          example: https://www.nextcentury.com/TA3/SIN_ID/TA3_TEAM_RUN_ID/TA2_TEAM_RUN_ID/TA1_TEAM_RUN_ID
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                fileName:
                  type: string
                  format: binary
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InlineResponse200"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
    delete:
      tags:
        - Blazegraph
      summary: Drop Blazegraph graph by URI
      description: Drop Blazegraph graph by URI
      operationId: graphDelete
      parameters:
        - name: graphURI
          in: query
          description: URI of graph
          required: true
          schema:
            type: string
          example: https://www.nextcentury.com/TA3/SIN_ID/TA3_TEAM_RUN_ID/TA2_TEAM_RUN_ID/TA1_TEAM_RUN_ID
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InlineResponse200"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
    put:
      tags:
        - Blazegraph
      summary: Update Blazegraph graph by URI
      description: Currently only supports rename of graph by URI
      operationId: graphRename
      parameters:
        - name: graphURIOrig
          in: query
          description: URI of OLD graph
          required: true
          schema:
            type: string
        - name: graphURINew
          in: query
          description: URI of NEW graph
          required: true
          schema:
            type: string
          example: https://www.nextcentury.com/TA3/SIN_ID/TA3_TEAM_RUN_ID/TA2_TEAM_RUN_ID/TA1_TEAM_RUN_ID
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InlineResponse200"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
  /graph/batch:
    post:
      tags:
        - Blazegraph
      summary: Batch upload of compressed TTLs and Import into Blazegraph
      description: Batch upload of compressed TTLs and Import into Blazegraph
      operationId: graphBatchUpload
      parameters:
        - name: graphURI
          in: query
          description: URI of graph
          required: true
          schema:
            type: string
          example: https://www.nextcentury.com/TA3/TA3_TEAM_RUN_ID/TA2_TEAM_RUN_ID/TA1_TEAM_RUN_ID
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                compressedFile:
                  type: string
                  format: binary
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InlineResponse200"
        "400":
          description: Invalid parameters
        "404":
          description: API not found
    delete:
      tags:
        - Blazegraph
      summary: Drop multiple Blazegraph graphs by base URI
      description: Drop multiple Blazegraph graph by base URI
      operationId: graphBatchDelete
      parameters:
        - name: graphBaseURI
          in: query
          description: Base URI of graphs
          required: true
          schema:
            type: string
          example: https://www.nextcentury.com/TA3
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InlineResponse200"
        "400":
          description: Invalid parameters
        "404":
          description: API not found          
...
components:
  schemas:
...
    graphs:
      type: array
      items:
        type: string
      example:
        - https://www.nextcentury.com/TA2/NCC/LDC
        - https://www.nextcentury.com/TA2/GAIA/M18_GAIA_1
...
    InlineResponse200:
      type: object
      properties:
        message:
          type: string
...